requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==2.0.1
mariadb==1.1.10         # MariaDB Connector/Python
python-dotenv==1.0.1
pydantic==2.9.2
streamlit==1.38.0
httpx==0.27.2
# Gemini (Google) embeddings; optional but recommended
google-generativeai==0.7.2